Revert "Fix int to double in Transaction Item price type. Fix web.xml routes to"
This reverts commit 5aab974311e66628f4bdddb5ed9eaf765fb5443d.

+++ src/java/model/TransactionItemDTO.java
@@ -13,7 +13,7 @@ public class TransactionItemDTO {
     private int userTransactionId;
     private int foodId;
     private int quantity;
-    private Double price;
+    private int price;
 
     // Getters and Setters
     public int getTransactionItemId() {
@@ -48,11 +48,11 @@ public class TransactionItemDTO {
         this.quantity = quantity;
     }
 
-    public Double getPrice() {
+    public int getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
 }
